LayoutTests/accessibility/aria-sort-changed-notification-expected.txt

This tests that changing the aria-sort value results in a SortDirectionChanged notification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axColumnHeader.sortDirection is 'AXAscendingSortDirection'
Toggling aria-sort
AXSortDirectionChanged notification for Account
PASS axColumnHeader.sortDirection is 'AXDescendingSortDirection'
Setting aria-sort to a random value
AXSortDirectionChanged notification for Account
PASS axColumnHeader.sortDirection is 'AXUnknownSortDirection'
Toggling aria-sort
AXSortDirectionChanged notification for Account
PASS axColumnHeader.sortDirection is 'AXAscendingSortDirection'
PASS notificationCount is 3
PASS successfullyParsed is true

TEST COMPLETE
